+++ src/time_evolution/brmesolve.jl
@@ -28,8 +28,8 @@ The return depends on `fock_basis`.
 """
 function bloch_redfield_tensor(
         H::QuantumObject{Operator},
-        a_ops::Union{Nothing, VectorOrTuple{<:QuantumObject{Operator}}},
-        c_ops::Union{Nothing, VectorOrTuple{<:QuantumObject{Operator}}} = nothing;
+        a_ops::Union{Nothing, AbstractVector, Tuple},
+        c_ops::Union{Nothing, AbstractVector, Tuple} = nothing;
         sec_cutoff::Real = 0.1,
         fock_basis::Union{Val, Bool} = Val(false),
     )
@@ -213,10 +213,10 @@ function brmesolve(
         H::QuantumObject{Operator},
         ψ0::QuantumObject,
         tlist::AbstractVector,
-        a_ops::Union{Nothing, VectorOrTuple{<:QuantumObject{Operator}}},
-        c_ops::Union{Nothing, VectorOrTuple{<:QuantumObject{Operator}}} = nothing;
+        a_ops::Union{Nothing, AbstractVector, Tuple},
+        c_ops::Union{Nothing, AbstractVector, Tuple} = nothing;
         sec_cutoff::Real = 0.1,
-        e_ops::Union{Nothing, VectorOrTuple{<:QuantumObject{Operator}}} = nothing,
+        e_ops::Union{Nothing, AbstractVector, Tuple} = nothing,
         kwargs...,
     )
     R = bloch_redfield_tensor(H, a_ops, c_ops; sec_cutoff = sec_cutoff, fock_basis = Val(true))